requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langfuse>=2.50.0
streamlit>=1.38.0
python-dotenv>=1.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
